refactor: 디자인 시스템 토큰 패키지 이동 및 네이밍 수정
