+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/designsystem/icon/MnMediumIcon.kt
@@ -1,6 +1,5 @@
 package com.pomonyang.mohanyang.presentation.designsystem.icon
 
-import IconSize
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Icon
@@ -8,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import com.pomonyang.mohanyang.presentation.designsystem.token.MnIconSize
 
 @Composable
 fun MnMediumIcon(
@@ -19,7 +19,7 @@ fun MnMediumIcon(
     Icon(
         painter = painterResource(id = resourceId),
         contentDescription = contentDescription,
-        modifier = modifier.size(IconSize.medium),
+        modifier = modifier.size(MnIconSize.medium),
         tint = tint
     )
 }
