+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/designsystem/token/MnStroke.kt
@@ -1,8 +1,8 @@
-package com.pomonyang.mohanyang.presentation.theme
+package com.pomonyang.mohanyang.presentation.designsystem.token
 
 import androidx.compose.ui.unit.dp
 
-object MhStroke {
+object MnStroke {
     val small = 0.5.dp
     val medium = 1.dp
     val large = 2.dp
